+++ plugin/src/main/kotlin/de/infolektuell/gradle/jextract/tasks/ExtractTask.kt
@@ -1,4 +1,5 @@
 package de.infolektuell.gradle.jextract.tasks
+
 import org.gradle.api.DefaultTask
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.file.RegularFileProperty
@@ -7,8 +8,10 @@ import org.gradle.api.file.FileSystemOperations
 import org.gradle.api.tasks.InputFile
 import org.gradle.api.tasks.OutputDirectory
 import org.gradle.api.tasks.TaskAction
+import org.gradle.work.DisableCachingByDefault
 import javax.inject.Inject
 
+@DisableCachingByDefault(because = "Extracting an archive is not worth caching")
 abstract class ExtractTask @Inject constructor(private val fileSystem: FileSystemOperations, private val archives: ArchiveOperations) : DefaultTask() {
     @get:InputFile
     abstract val source: RegularFileProperty